server_python/blog/pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["zyb <1808639364@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
loguru = "^0.7.2"
redis = "^5.1.1"
djangorestframework-simplejwt = "^5.5.0"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
